backend/requirements.txt
sentry-sdk
python-dotenv
flask_cors
flask
werkzeug
flask_restplus
flask_jwt_extended
flask_bcrypt
flask_migrate
flask_redis
mockredis
sqlalchemy_utils
pymysql
flask-migrate
flask_script
flask_mail
email_validator
flask_jwt_extended
requests